requirements.txt
asgiref==3.8.1
bleach==6.2.0
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.0
webencodings==0.5.1
whitenoise==6.9.0
